pyproject.toml
[tool.poetry]
name = "cindocubot"
version = "0.1.0"
description = ""
authors = ["Fennr <fenrir1121@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
disnake = "^2.7.0"
PyYAML = "^6.0"
peewee = "^3.15.3"
aiohttp = "^3.8.3"
python-i18n = {git = "https://github.com/Krutyi-4el/python-i18n.git", branch="master"}
pytz = "^2022.5"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"
python-dateutil = "^2.9.0.post0"
dateparser = "^1.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
prospector = "^1.9.0"
autopep8 = "^2.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 101
disable = "too-few-public-methods"